autoconf: use pkg-config to detect SQLITE3
Before this fix, ./configure used once code from cmulocal/sql.m4 to
find out, if libsqlite3 is installed on the system and turned sqlite3
support on, only when --with-sqlite3 was provided.

At the same time, --enable-http used pkg-config to detect libsqlite3, so
that it was possible to build http (with sqlite3), but without having
sqlite3 support in cyrus.

The furhter, building cyrus with sqlite3 is not possible, because imap/user.c
depends on "lib/sqldb.h", which depends on <sqlite3.h>.

This patch:
 -- makes consistent use of pkg-config to detect sqlite3 support;
 -- enables sqlite3 by default, if pkg-config finds it;
 -- rejects the combination --enable-http --without-sqlite3;
 -- deletes sqlite3 detection code from cmulocal/sql.m4.